requirements.txt
fastapi>=0.115.0,<1.0.0
uvicorn>=0.30.6,<1.0.0
python-dotenv>=1.0.1,<2.0.0
pydantic>=2.10.0,<3.0.0
openai>=1.51.2,<2.0.0
chromadb>=0.5.5,<1.0.0
sentence-transformers>=3.1.1,<6.0.0
requests>=2.32.3,<3.0.0
python-multipart>=0.0.9,<1.0.0
Pillow>=10.4.0,<13.0.0
numpy==1.26.4; python_version < "3.13"
numpy>=2.2.0; python_version >= "3.13"
